model_training/requirements.txt
Pillow
numpy
boto3
torch
torchvision
pandas
onnx